+++ script/popup/render-room-filters.js
@@ -3,5 +3,5 @@
   const ROOM_BOOKING_FILTERS_UI_GROUP = 'room-booking-filter-ui-group';
 
   const uiGroupWrapper = document.getElementById(ROOM_BOOKING_FILTERS_UI_GROUP);
-  uiGroupWrapper.appendChild(await asyncRenderRoomBookingFilters(persistPair));
+  uiGroupWrapper.appendChild(await asyncRenderRoomBookingFilters(persistPairSync));
 })();