+++ script/popup/set-feature-toggles.js
@@ -5,7 +5,7 @@
   // Restores feature toggle values using the preferences stored in chrome.storage
   async function restoreOptions() {
     // fill out saved feature toggles. use default if nothing is found
-    const settings = await getFromStorage(DEFAULT_FEATURE_TOGGLES);
+    const settings = await getFromStorageSync(DEFAULT_FEATURE_TOGGLES);
     Object.keys(DEFAULT_FEATURE_TOGGLES).forEach(key => document.getElementById(key).checked = settings[key])
   }
 
@@ -13,6 +13,6 @@
 
   // add feature toggle click listener
   Object.keys(DEFAULT_FEATURE_TOGGLES).forEach(key =>
-    document.getElementById(key).addEventListener("click", e => persistPair(key, e.target.checked))
+    document.getElementById(key).addEventListener("click", e => persistPairSync(key, e.target.checked))
   );
 })();