+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## 2.2.0
+- [feature] Support searching room for recurring meetings
 - [enhancement] Show a "paused" sign on the icon when the worker is stopped
 
 ## 2.1.0